+++ src/front/js/component/navbar_home.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import { Buscador } from "./navbar_home";
+
 
 export const Navbar = () => {
 	return (
@@ -10,6 +10,13 @@ export const Navbar = () => {
 					<Link to="/"><img className="img-fluid logo" src="https://i.postimg.cc/q77XXd3d/Young-Star-removebg-preview-adobe-express.png" alt="click_logo"/>
 					</Link>
 				</div>
+
+				<div className="d-flex">
+					<input className="buscador form-control me-2" type="search" placeholder=" Buscador" aria-label="Search">
+					</input>
+					<button className="btn btn-primary mx-1"><i className="fas fa-search me-2 pt-2"></i></button>
+				</div>
+
 				<div className="d-flex flex-row-reverse">
 					<Link to="/signup">
 						<button className="btn btn-primary mx-1">Signup</button>
