+++ src/front/styles/home.css
@@ -1,13 +1,26 @@
 /*
     home.css: This website contains selectors only used in home.css
 
-    All pages share the styles on index.css but you should create 
-    one more css for each page that will contain the selected used 
+    All pages share the styles on index.css but you should create
+    one more css for each page that will contain the selected used
     on that page only (the ones not reused in other pages).
 */
+body {
+  min-height: 100%;
+}
+
+
+.footer { position:relative; height: 30px; bottom: 0; width: 100%;
+
+}
 
 .buscador {
   width: 30rem;
+}
+
+.logo {
+  width: 12rem;
+  height: 4rem;
 }
 
 .profile-photo {
